Disabled nonterminal referencing across files
as in Epsilon every file is its own global scope or, said otherwise, in
Eclipse the local scope is equal to the global scope.
So, an IGlobalScope provider has been implemented which returns the
IScope.NULL_SCOPE and has been registered.
Additionally, the SimpleLocalScopeProvider has been registered replacing
the default scope implementation as this is sufficient for the Epsilon
language.