Migrated node dependency 'vscode' to '@types/vscode'.
In fact the old version have broken the build due to not availabe API
server for downloading vscode.d.ts.
While trying to update to a newer version of vscode it was figure
out that the dependency structure had bee significantly changed.
So, the vscode dependency was migrated according to
https://code.visualstudio.com/api/working-with-extensions/testing-extension#migrating-from-vscode