Implemented a combined Preference/Property page
templated from org.eclipse.xtext.builder.preferences.
This allows to make project specific settings as known from other
Eclipse settings.
However, the configuration is not connected yet to the Validator and
Executor. This step needs still to be done by using providing a similar
implementation as the EpsilonPropertyProvider. .. The preference store
is another one and needs to be based on the language name.