Let store node be the language name and preference keys full qualified.
That means, for project properties and preferences the same keys are
used at the workspace preference store and at the properties project
store.
The properties are read out at Executor via the PropertiesProvider as
before, a static implementation. All works well, the workspace
preferences changes as well as the project properties changes are
reflected in the Executor.
The next step is to concentrate all in the EpsilonPreferences and make
it non-static to get it injected into the Executor or Evaluator.